Move git::ext and internal::result to separate crates
