+++ git-ext/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "radicle-git-ext"
+version = "0.1.0"
+authors = ["The Radicle Team <dev@radicle.xyz>"]
+edition = "2018"
+license = "GPL-3.0-or-later"
+description = "Utilities and extensions to the git2 crate"
+
+[dependencies]
+multibase = "0"
+multihash = "0"
+percent-encoding = "2"
+thiserror = "1"
+
+[dependencies.git2]
+version = ">= 0.13.12"
+default-features = false
+features = []
+
+[dependencies.serde]
+version = "1"
+features = ["derive"]
+
+[dependencies.radicle-std-ext]
+path = "../std-ext"
+
+[dev-dependencies]
+anyhow = "1"
+assert_matches = "1"
+lazy_static = "1"
+pretty_assertions = "0"
+regex = "1"
+serde_json = "1"
+
+[dev-dependencies.librad-test]
+path = "../librad-test"
