+++ git-ext/src/reference/name.rs
@@ -189,7 +189,7 @@ impl From<RefLike> for PathBuf {
 ///
 /// ```rust
 /// use std::{convert::TryFrom, path::Path};
-/// use librad::git::ext::reference::name::*;
+/// use radicle_git_ext::reference::name::*;
 ///
 /// assert_eq!(
 ///     &*OneLevel::from(RefLike::try_from("refs/heads/next").unwrap()),
@@ -274,7 +274,7 @@ impl From<OneLevel> for PathBuf {
 ///
 /// ```rust
 /// use std::{convert::TryFrom, path::Path};
-/// use librad::git::ext::reference::name::*;
+/// use radicle_git_ext::reference::name::*;
 ///
 /// assert_eq!(
 ///     &*Qualified::from(RefLike::try_from("laplace").unwrap()),
