+++ librad/src/git.rs
@@ -15,7 +15,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod ext;
 pub mod include;
 pub mod local;
 pub mod p2p;
