+++ librad/src/git/storage/config.rs
@@ -19,13 +19,12 @@
 
 use std::{convert::TryFrom, io};
 
+use keystore::sign;
+use radicle_git_ext::is_not_found_err;
+use radicle_std_ext::result::ResultExt as _;
 use thiserror::Error;
 
-use keystore::sign;
-
 use crate::{
-    git::ext::is_not_found_err,
-    internal::result::ResultExt,
     keys::SecretKey,
     meta::{
         entity::{Signatory, Verified},
