+++ librad/src/identities/sealed.rs
@@ -15,8 +15,8 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::git::ext;
+use radicle_git_ext::Oid;
 
 pub trait Sealed {}
 
-impl Sealed for ext::Oid {}
+impl Sealed for Oid {}
