+++ std-ext/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "radicle-std-ext"
+version = "0.1.0"
+authors = ["The Radicle Team <dev@radicle.xyz>"]
+edition = "2018"
+license = "GPL-3.0-or-later"
+description = "Monkey patches of std types"
